kern_descrip: fget_only_user is the one fd accessor without the cast
Reading the 38 sites the parameter fix made newly visible to
onesided_index.py.  Three were in the descriptor table, which is where
a signed index matters most.  All three turned out to be bounded --
and two of the three are bounded by an idiom the rule did not know,
which is worth more than the readings.

  sys/kern/kern_descrip.c fget_only_user()

    if (__predict_false(fd >= fdp->fd_nfiles))
            return (EBADF);

fdt_nfiles is `int' (filedesc.h:79), so this is a SIGNED comparison
and a negative fd walks straight past it into `fdt_ofiles[fd]'.  The
three inline accessors beside it -- filedesc.h:317, :330 and :343 --
all write `(u_int)fd >= (u_int)fdp->fd_nfiles', where the cast makes
one comparison reject both ends.  This is the one in the family that
does not.

NOT A LIVE BUG, and the commit should not be read as claiming one.
Both callers come through select and poll, and pollscan()
(sys_generic.c) skips `fds->fd < 0' before an entry is ever registered,
so pollrescan() cannot walk one.  That is a guard two functions away
in another file, protecting the fast path into the descriptor table,
and the cast costs nothing.  Both definitions -- CAPABILITIES and the
#else twin -- now carry it.

  the two that were already right, and why they read as wrong

close_range_flags() and close_range_impl() take `u_int lowfd, u_int
highfd' and copy into `int fd'.  The guard is `if (fd > highfd) goto
out', and because highfd is unsigned the comparison converts fd to
unsigned too -- so a lowfd above INT_MAX becomes a huge unsigned,
fails the test, and leaves before any subscript.  Correct by an
implicit conversion.  Tabled rather than touched.

  AND THE RULE LEARNED THE CAST

fgetvp_lookup_smr() spells the bound explicitly, `(u_int)fd >=
fdt->fdt_nfiles', and onesided_index.py called it one-sided because
_floor_re() looks for a comparison against zero or -1.  An unsigned
cast on V IS a floor -- it is the whole point of the idiom -- so it is
one now.  The cast has to be to an unsigned type and has to be on V:
`(int)V' is not a floor, and neither is a cast on the other operand.

  ONESIDED  302 -> 288 advisory.  Fourteen sites that were never
            one-sided, and the kernel's commonest way of saying so.
  MOD       0 not on the record, unchanged.  The gate stays green.

Four tests on the cast form, both directions.  One marker pair added
and revert-verified; the kern_descrip.c entry merged into the existing
one rather than added beside it, which check_pbsd_marks.py catches as
a duplicate key -- a repeated key silently replaces the first, and
every marker in it.

Measured: sys/kern ERROR unchanged, all 5 on the record.

+++ hbsd/src/sys/kern/kern_descrip.c
@@ -3614,7 +3614,23 @@ fget_only_user(struct filedesc *fdp, int fd, const cap_rights_t *needrightsp,
 	MPASS(FILEDESC_IS_ONLY_USER(fdp));
 
 	*fpp = NULL;
-	if (__predict_false(fd >= fdp->fd_nfiles))
+	/*
+	 * PBSD: bound both ends, as the rest of the family does.
+	 * fdt_nfiles is an int (filedesc.h:79), so `fd >= fd_nfiles'
+	 * is a SIGNED comparison and a negative fd walks straight
+	 * past it into fdt_ofiles[fd].  The three inline accessors
+	 * beside it -- filedesc.h:317, :330 and :343 -- all write
+	 * `(u_int)fd >= (u_int)fdp->fd_nfiles', where the cast makes
+	 * one comparison reject both ends; this one did not.
+	 *
+	 * Not reachable today: both callers come through select and
+	 * poll, and pollscan() skips `fds->fd < 0' before an entry is
+	 * ever registered, so pollrescan() cannot walk one.  That is
+	 * a guard two functions away in another file, protecting the
+	 * fast path into the descriptor table.  The cast costs
+	 * nothing and says it here.
+	 */
+	if (__predict_false((u_int)fd >= (u_int)fdp->fd_nfiles))
 		return (EBADF);
 
 	fdt = fdp->fd_files;
@@ -3640,7 +3656,8 @@ fget_only_user(struct filedesc *fdp, int fd, const cap_rights_t *needrightsp,
 	MPASS(FILEDESC_IS_ONLY_USER(fdp));
 
 	*fpp = NULL;
-	if (__predict_false(fd >= fdp->fd_nfiles))
+	/* PBSD: the same cast, in the !CAPABILITIES twin. */
+	if (__predict_false((u_int)fd >= (u_int)fdp->fd_nfiles))
 		return (EBADF);
 
 	fp = fdp->fd_ofiles[fd].fde_file;
